Fix ZipSlip bug.
The unsanitized path of a zip archive entry, which may contain '..', was used
directly to resolve the destination path for the files being unzipped.

Extracting files from a malicious archive without validating that the
destination file path is within the destination directory can cause files
outside the destination directory to be overwritten.